feat: deprecate `hash base64` command
