+++ crates/nu-command/src/default_context.rs
@@ -382,7 +382,7 @@ pub fn create_default_context(cwd: impl AsRef<Path>) -> EngineState {
             Hash,
             HashMd5::default(),
             HashSha256::default(),
-            Base64,
+            HashBase64,
         };
 
         // Experimental
