+++ crates/nu-command/src/hash/mod.rs
@@ -1,10 +1,8 @@
-mod base64;
 mod generic_digest;
 mod hash_;
 mod md5;
 mod sha256;
 
-pub use self::base64::Base64;
 pub use self::hash_::Hash;
 pub use self::md5::HashMd5;
 pub use self::sha256::HashSha256;
